Make preview functionality work in Rails 6.0.0
Rails 6.0.0 introduced a check on the `formats` that you can render:

https://github.com/rails/rails/pull/35661

Since that commit, you can only specify the `formats` as symbols, not
strings. The render statement in the actionmailer controller was
updated in the PR as well:

https://github.com/rails/rails/pull/35661/files#diff-32907ca93d739460000
a3f554515caa4

This commit introduces the fix to our controller too. Without this, the
preview page will raise an error:

```
     ArgumentError:
       Invalid formats: "html"
     # ./lib/mail/notify/mailers_controller.rb:28:in
`render_preview_wrapper'
```