Bump notifications-ruby-client
This allows updating to 5.1.1, which fixes an issue that prevents the
exception message from being picked up by Sentry.

https://github.com/alphagov/notifications-ruby-client/blob/master/CHANGE
LOG.md#511

I’ve made the requirement looser so users can bump the gem to any
version that shouldn’t be breaking (< 6).

The breaking change between 4 and 5 is that the gem drops support for
Ruby 2.3.