Require `rails/mailers_controller` specifically
Running `rails zeitwerk:check` gives the following error:

```
NameError: uninitialized constant ApplicationHelper
```

I’ve tracked this down to the use of `Rails::MailersController.prepend`
in railtie.rb. A quick and dirty fix is to require
`rails/mailers_controller` specifically in the Railtie. This seems like
an upstream bug in Rails itself (especially as this is a somewhat
obscure part of the Rails ecosystem).