Fix specs
Inside both `Mailer.view_mail` and `Mailer.template_mail` the call to
`mail` is made with a complete hash, whilst this spec will pass in Ruby
2.x it will fail in 3.x.

You can see that:

`mail(headers.merge(template_id: template_id))`

By making this change it passes in both and better reflects the actual
behaviour.

This does not change the implementation so should not break anything.