Pin Rails to fix test suite
Rails 7.1 introduces a change to the email preview template that our
testing suite  cannot handle yet, a `link_to` causes the render to
fail.

To give us the space to fix the underlying issue, we'll pin Rails at
7.0.8, this is only for the test suite, the gem can still support
other versions of Rails as the issue will not be present in actual
installaions.

With the test suite working again we can work on some other issues,
update some dependancies and come back to this link_to problem.