WIP: Part 2
Continuing to delve into this.

Collected notes in the wiki on GH.

I think I have a better or at least different way to prepend the
mailer_controller preview method for our purposes.

We want to check if the email is going to be delivered by Notify and
render a different preview when that is the case.

The main blocker right now is not having a real Notify account to test
with - we call the api to generate a preview, which is what we want to
show in the Rails preview, but without a real API key I can't get a
handle on that.