Remove MailersController overrides
All this code relates to ActionMailer previews and it is what we think
causes the previews to fail in Rails > 7.0.8.

Prepending any objectm, as was done here has a high price and is not
needed for the previews to work in Rails.

v2 will use a different approach to this.