Add a previews controller
Right now, the simplest way to render the html body that comes back from
Notify in a layout so that the user gets the closet preview to the real
email is to use the `renderer` from a controller.

Whilst we could use ApplicationController, it feels a little brittle, so
instead we create a completely empty controller so we can access it's
renderer.

This will get used later to render previews.