Extend Mail::Message
We want to store the tempalte id, personalisation, reference and reply
to id options inside the `message` object, right now these are added to
the header of the message, which feels a little clunky.

Here we use the existing module to add the attributes to message so we
can set and get the values.

We also drop the `preview` method, this method hides the fact that the
impelmentation of the `preview` method is in the DeliveryMethod - on
balance the extra code to call the method seems worth the price of the
understanding where messages are being sent - to be honest, it already
feels odd that the message calls a method and passes it self - but that
might just be me?