Update the Mailer
The Mailer is what is used to provide applcations with Notify emails.

This class is used as the parent for new subclasses that provide the
application specific emails.

Like a module, we have to test this by creating a new TestMailer that
inherits from our updated class.

The main thing to notice here is that we have to prevent Rails from
carrying out the default behaviour of looking for views to provide the
parts of the email (the html and plain text content), we do so by
calling `mail` with a block that sets the parts to empty object - these
are then replaced later with the actual content.