Update delivery method
The delivery method is where the actual sending and in our case
previewing happens - for that reason it is where the calls to the Notify
API are made.

We've updated this class to make it work for the new way messages are
formed and for simplicity.

We've stopped using webmock in the tests as that felt too much
intergration, it felt like we are testing the notifications_client
rather than our code as long as we pass the correct params to the client
we should feel confident the right thing will happen.