Update dependencies
We want to support as many versions of Rails as we can, but anyone
running the gem in Rails 5 should really be on the last version.

As we start to update the gem to version 2 the latest release of the
notificaitons gem is also available, so we update that and all
subsequent minor and patch version should be applied.