Update CI workflow
We don't need to cache anything as the ruby/setup-ruby@v1 action takes
care of that for us.

It also sets up the appropriate version of Bundler for our Ruby version
(currently Ruby 2.7.8 and Bundler 2.4.22).

The sqlite setup is also taken care of for us.

Lastly some naming niceties, e.g this is our continuous intergration
rather than a 'build'.