Refactor blank_allowed
If passed a nil in the personalisation, the Notify API will raise an
error.

The `blank_allowed` method lets developers opt-in to allow a blank
value, here 'blank' probably means `nil`.

As v2 no longer relies on the `Personalisation` class to handle
personalisations we want to remove the class, but this method still
relies on it.

We want to keep the method in-case it is used.

In v2, we really only need to know when a developer is opting in by
calling the method and then convert nil into an empty string.