Add simple Notify Integration tests
We want to test the mailer previews, this requires some effort.

We also want to be able to test in different versions of Rails.

We use a customisable Docker image that installs a bare-bones Rails app
and adds the mail-notify gem.

We then copy over some pre-made system tests and run them.

There are a lot of things to consider in this work:

- we want to support older rubies, 2.7.8 right now
- we have to integrate with the real Notify API to fetch previews
- the Notify account, API key and template IDs are tightly coupled

With this now working locally we can move it to GitHub actions.

This work will support Rails 6 and 7, unfortunately 5 will require more
work and may not be worthwhile.