Integration test for Rails 5
This is the extra work required to get our simple integration test
running in Rails 5.

It boils down to remove the assumed reliance on Chrome for system tests
that we do not use anyway, we use the rack_test driver for ease.

Please note, these test are expected to FAIL on the 1x branch, they will
only pass on the new v2 branch.