Update integration GitHub Action
We added the new sending integration tests so we want to runs those.
Because we know the previewing and the sending are the two main
features, we want them to show up in GitHub separately so we break the
build and test jobs up passing the built image between via uploading as
an artifact.

There is nothing secret in the built image as we only have on secret and
that is passed into the running container not the image.

Until v2 is released, we expect the preview tests to FAIL but the
sending should work as this reflects the current state of the gem.