Fix builds for Rails integration tests
A couple of things changed recently that broke these for us:

- Rails 7.2 beta dropped support for Ruby < 3.1
- ffi dropped support for Rubygems < 3.3.22

We have updated Rubygems and bundler to new enough versions to
compile the ffi gem.

We also had to add the build-essential package into a later build step
as ffi seems to get built a second time - this is possibly incorrect and
will need looking into.

We should possibly decide to at least drop testing for end of life versions:

- Ruby 2.7.8
- Rails 5.2.8.1