Refactor mailer methods
We only keep a Mail::Message for use in previews, it is never actually
sent, email sending is handled by Notify, not the host application.

The Notify API does not accept additional custom headers, so passing
custom headers to the message via the call to `mail` serves no purpose
as those headers will never be used (other than in the previews, where
they get shown).

This work refactors the two mailer methods so that only the `to` and
`subject` headers are set on the Mail::Message returned.