Change how we specify the gem branch
We install the gem to run our integration tests. They recently started
to fail and the issues seemed to be that bundler couldn't locate the
branch.

The branch name is supplied by `github.ref` switching to the logic here
seems to have fixed this.

`head_ref` for PRs with a fall back to `ref_name` if it is not a PR.

Documentation:

https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/accessing-contextual-information-about-workflow-runs#github-context