Run integration tests on `pull_request_target`
To run the integration tests, we need the `NOTIFY_API_KEY` secret
to be present, running on `pull_request_target` ensures we run
the action in the context of this repo (albeit the `main` branch),
pulling down the changes from the repo. This means that there is
a small security risk in that a bad actor could reveal our secrets
using exploits in their code (see https://securitylab.github.com/resources/github-actions-preventing-pwn-requests/)
but given that we have to explicitly allow actions to run when coming
from external contributors, this is a small risk.

Additionally, we’ll add a CodeQL action to carry out static analysis
on any submitted code, reducing the risk even further.