Run integration tests after workflow run
It turns out that `pull_request_target` runs regardless of permissions,
as code that is run is considered to be trusted, which we don’t
necessarily want.

Changing this to `workflow_run` ensures that the action is ONLY run
after the rest of the CI runs (which will need to be accepted for
external contributions). Like `pull_request_target`, `workflow_run`
triggers have access to secrets, but won’t get triggered until the
other actions are approved and have run.