Run integration tests safely
This tweaks the integration tests to run on `pull_request_target`.

Because `pull_request_target` actions run regardless of if the
repo owner has approved them, we first check if the person who
triggered the run has admin permissions on the repo. If they don’t,
then the action exits early and fails.

Then once the code has been reviewed and the other actions have been
approved, an admin user can rerun the workflow with the correct
permissions.